requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
langchain>=0.1.14,<0.2
langchain-community>=0.0.30
openai>=1.10.0,<2.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
pydantic>=2.5
python-multipart==0.0.6
requests>=2.31
aiohttp>=3.9
python-dotenv==1.0.0aiohttp==3.9.1
aiohttp==3.9.1
openai==1.58.1